Enqueue period_bytes during playback
This commit enqueues requests of period_size bytes instead of individual
descriptors. Each request is a set of pointers to descriptors that
contain the payload. This commit modifies consume() to be able to read
through the descriptors of a given request. Thus, consume() can read up
to period_bytes from a request.

Signed-off-by: Matias Ezequiel Vara Larsen <mvaralar@redhat.com>