sound: soft-error on invalid PCM state transitions
Continuing the argument from the previous commit, "the spec does not state
rejecting a PCM life-cycle control request if the transition is not
possible. (The possible state transitions are non-normative
statements.)".

Instead, log the error but still follow the guest's request as far as
possible. The controlq response should not return VIRTIO_SND_S_BAD_MSG
if the message parameters (i.e. the stream ID) are correct. Instead it
can return VIRTIO_SND_S_IO_ERR, though not necessarily.

Signed-off-by: Manos Pitsidianakis <manos.pitsidianakis@linaro.org>